fix(Parser & GUI): rename respective classes
